+++ src/server/m_tcp_server.js
@@ -6,7 +6,8 @@
 */
 
 const ModbusSlave = require('../protocol/modbus_slave');
-const netServer = require('../net/tcpserver');
+const tcpServer = require('../net/tcpserver');
+const udpServer = require('../net/udpserver');
 const Request = require('../protocol/request');
 const Response = require('../protocol/response');
 
@@ -20,9 +21,18 @@ class ModbusTCPServer extends ModbusSlave {
   /**
   * Create a Modbus Tcp Server.
   * @param {number} p Port to listen.
+  * @param {string} tp. Transport layer. Can be tcp, udp4 or udp6
   */
-    constructor(p=502){
+    constructor(p=502, tp = 'tcp'){
       super();
+
+      var transportProtocol
+      if(typeof tp == 'string'){
+        transportProtocol = tp
+      }
+      else{
+        throw new TypeError('transport protocol should be a string')
+      }
 
       var self = this;
 
@@ -30,7 +40,18 @@ class ModbusTCPServer extends ModbusSlave {
       * network layer
       * @type {object}
       */
-      this.netServer = new netServer(p);
+     switch(transportProtocol){
+
+        case 'udp4':
+          this.netServer = new udpServer(p, 'udp4');
+          break;
+        case 'udp6':
+          this.netServer = new udpServer(p, 'udp6');
+          break;
+        default:
+            this.netServer = new tcpServer(p);
+     }
+      
 
       //Adding listeners to netServer events
       this.netServer.onData = this.ProcessModbusIndication.bind(this);
@@ -195,7 +216,7 @@ class ModbusTCPServer extends ModbusSlave {
     */
     ProcessModbusIndication(connectionID, dataFrame){
       var self = this;
-      let socket = this.netServer.activeConections[connectionID];
+      let socket = this.netServer.activeConnections[connectionID];
       
 
       /**